requirements.txt
# Core parsing
PyPDF2==3.0.1
python-docx==1.1.0
python-docx==1.1.0
pymupdf==1.27.2

# LLM
openai==1.30.1
httpx==0.27.2

# Utilities
pydantic==2.7.1
typing-extensions==4.11.0
python-dotenv==1.1.0

# Testing
pytest==8.2.1